refactor(sdk): [wip] use ContentService
